loadtest/requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
dill==0.2.8.2
fastai==1.0.20
fastprogress==0.1.15
Flask==1.0.2
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
humanfriendly==4.17
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
locustio==0.9.0
MarkupSafe==1.1.0
matplotlib==3.0.1
mccabe==0.6.1
molten==0.7.1
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
mypy-extensions==0.4.1
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
plac==0.9.6
pluggy==0.8.0
preshed==2.0.1
prometheus-client==0.4.2
py==1.7.0
pylint==2.1.1
pyparsing==2.3.0
pytest==3.10.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
regex==2018.11.3
requests==2.20.0
scipy==1.1.0
six==1.11.0
spacy==2.0.16
thinc==6.12.0
toolz==0.9.0
torch-nightly==1.0.0.dev20181103
torchvision-nightly==0.2.1
tqdm==4.28.1
typing==3.6.6
typing-extensions==3.6.6
typing-inspect==0.3.1
ujson==1.35
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
wsgicors==0.7.0
